delete meetings.js, replace with calendar and event files for both controllers and routes
