+++ routes/index.js
@@ -1,33 +1,34 @@
 import { Router } from 'express';
 
-import notificationRoutes from './user/notifications.js';
-import ticketsRoutes from './tickets.js';
 import authRoutes from './user/auth.js';
-import userRoutes from './user/users.js';
-
-import projectRoutes from './project/projects.js';
-import taskRoutes from './project/tasks.js';
 import availabilityRoutes from './calendar/availability.js';
-import meetingRoutes from './calendar/meetings.js';
+import calendarRoutes from './calendar/calendars.js';
 import chatRoutes from './chats.js';
-import privateChatRoutes from './chat/privateChat.js';
+import eventRoutes from './calendar/events.js';
 import groupChatRoutes from './chat/groupChat.js';
 import mediaRoutes from './chat/media.js';
+import notificationRoutes from './user/notifications.js';
+import privateChatRoutes from './chat/privateChat.js';
+import projectRoutes from './project/projects.js';
+import taskRoutes from './project/tasks.js';
+import ticketsRoutes from './tickets.js';
+import userRoutes from './user/users.js';
 const router = Router();
 
 // would it be better to append "/user" or "/chat" etc. to the routes here instead?
 router.get('/', (req, res) => res.send('api root'));
 router.use('/', authRoutes);
-router.use('/', userRoutes);
-router.use('/', projectRoutes);
-router.use('/', notificationRoutes);
-router.use('/', ticketsRoutes);
 router.use('/', availabilityRoutes);
+router.use('/', calendarRoutes);
 router.use('/', chatRoutes);
-router.use('/', privateChatRoutes);
+router.use('/', eventRoutes);
 router.use('/', groupChatRoutes);
 router.use('/', mediaRoutes);
-router.use('/', meetingRoutes);
+router.use('/', notificationRoutes);
+router.use('/', privateChatRoutes);
+router.use('/', projectRoutes);
 router.use('/', taskRoutes);
+router.use('/', ticketsRoutes);
+router.use('/', userRoutes);
 
 export default router;
